Rename main state to `RunningState` and move to defender/state.rs
